+++ includes/matrix.h
@@ -6,7 +6,7 @@
 /*   By: llai <llai@student.42london.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/29 16:45:55 by llai              #+#    #+#             */
-/*   Updated: 2024/05/02 15:42:39 by llai             ###   ########.fr       */
+/*   Updated: 2024/05/03 17:11:56 by llai             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ typedef struct s_matrix
 {
 	int		rows;
 	int		cols;
-	double	**data;
+	float	**data;
 }	t_matrix;
 
 t_matrix	create_matrix(int rows, int cols);
@@ -32,17 +32,17 @@ t_tuple		matrix_tuple_multiply(t_matrix A, t_tuple b);
 t_matrix	identity_matrix(t_matrix m);
 t_matrix	init_identitymatrix(int size);
 t_matrix	transpose(t_matrix A);
-double		determinant(t_matrix M);
+float		determinant(t_matrix M);
 t_matrix	submatrix(t_matrix A, int rowToRemove, int colToRemove);
-double		minor(t_matrix A, int row, int col);
-double		cofactor(t_matrix A, int row, int col);
+float		minor(t_matrix A, int row, int col);
+float		cofactor(t_matrix A, int row, int col);
 t_matrix	inverse(t_matrix M);
 
 // matrix_transform.c
-t_matrix	translation(double x, double y, double z);
-t_matrix	scaling(double x, double y, double z);
-t_matrix	rotation_x(double theta);
-t_matrix	rotation_y(double theta);
-t_matrix	rotation_z(double theta);
-t_matrix	shearing(double xy, double xz, double yx, double yz, double zx, double zy);
+t_matrix	translation(float x, float y, float z);
+t_matrix	scaling(float x, float y, float z);
+t_matrix	rotation_x(float theta);
+t_matrix	rotation_y(float theta);
+t_matrix	rotation_z(float theta);
+t_matrix	shearing(float xy, float xz, float yx, float yz, float zx, float zy);
 #endif // !MATRIX_H
